presenter/github: Display only commit summary (shorter message).
This is more consistent with code.google.com presenter, and in many
cases more readable.
It would be nice to present the full commit message (with details), but
it's better to do that once it can be displayed nicely hidden with a
way to expand/hide it (similar to how GitHub presents them). Until
then, it's better to just display an easier to read summary and rely on
the direct commit link for more details.