Improve reliability of starting http server and opening browser client.
Previously there was a race condition in that the browser was opened
first and would start navigating to the target URL while the http
server would hopefully be started in time. To remove that race
condition and make startup more reliable, start listening first, open
browser client second.
Perform Open in a background goroutine in case it blocks execution on
some systems, which would prevent starting the http server from being
started.
Should help with second issue in #26.