presenter: Remove "code.google.com/..." presenter.
It was no longer being used since Go and its subrepos have moved from
code.google.com to GitHub.
It is currently failing to compile because of upstream API changes. The
upstream API may have more changes, and this code is unused, so it
makes more sense to remove it.
I may still want to bring it back to support arbitrary package
presenters (i.e., those not on GitHub), but that's a separate
enhancement.
Fixes #37.