Don't update dependencies of repos.
When updating repo A.com/b/c/..., only update the given repo and not
all of its dependencies. This fixes #39.

It's still possible to update existing dependencies individually. But
it won't fetch new repos, which may or may not be desired. User can
always go get new repos when needed.