Port frontend JavaScript code to Go.
With this change, this application will be written completely in Go, both
in the backend and in the frontend. Coincidentally, this is the last
hand-written JavaScript code I have in any of my personal repositories.
After removing it, I will not have any JavaScript left over. I prefer
working with Go code (and I think many others do too).