repo: Use govendor fetch command to update.
It's a better idea since it fetches directly from the source, rather
than local disk. Local disk may not have the same version of the
package, or no package at all.

Also specify version to be fetched. In most cases it'll be the same as
what GPS showed earlier, but it prevents potential race conditions if
there was a newer commit since you've last checked for updates.