Improve internal documentation of concurrent processing.
The gophers do a lot of cool and hard work to power Go Package Store
efficiently, so give them some recognition by documenting what they do.

Add missing package and exported symbol docs.

Rename Presented to RepoPresenters. The goal is to make it more clear
what it does and when it can be used.