Rename presenter to presentation; provider to presenter.
These new names make a lot more sense and better correspond to reality.
A presenter is something that takes a repository, performs the work,
and returns a presentation of that repository. The presentation is
ready to be displayed and does no work.

Clean up, simplify, improve organization and style of packages.
Avoid stutter in many names.