Render HTML on frontend.
This is a large change that primarily moves HTML rendering and display
logic from backend to frontend (#67).

Previously, the HTML for displaying updates was rendered on backend and
streamed to browser. This worked surprisingly well and got me far, but
in order to be able to have more fine grained control over frontend
details, it was no longer viable to keep doing that. Now, the HTML is
fully rendered on frontend, and most of the logic resides on the
frontend. The backend provides services to the frontend. See issue #67
for full rationale why this is desired.

Implement a long-standing feature request of having an "Update All"
button (#6). This is both made possible and easy thanks to the frontend
HTML rendering.

This change partially helps #63, but also enables potential future
changes to help it even more.

In general, the move to frontend HTML rendering will help potentially
achieve some of enhancements described in #8.

Close #66 by removing the popup altogether. It wasn't well implemented,
so it's better to remove. In the future, a better replacement
implementation of the notification (without the modal popup) can be
considered.

Resolves #67.
Closes #66.
Helps #63.
Helps #8.
Resolves #6.