workspace: Check local branch matches default branch first.
We want to check if local branch doesn't match default (as determined
from remote) branch before we check if local revision is empty. This
way, we can report a better error message in a case where the default
branch doesn't exist locally.

Add a sanity check to ensure Remote.Branch is not empty. It's not
expected to be. For consistency with Remote.Revision being empty check.

Fixes #74.