workspace: Detect and skip updates when branches have diverged.
When a repository has local commits that are ahead of remote, but
there's no remote update, don't print anything. It's not a problem if
there are no updates.

When that's the case and there are updates (i.e., branches have
diverged), then skip presenting such updates, since they would fail to
apply anyway (and print reason to stderr, as is the case for other
reasons updates can't be applied/displayed).

Update internal documentation to accurately describe reality. Previous
version was not accurate.