cmd/Go-Package-Store: Add initial naive dep support via -dep flag.
This is the first naive implementation. It reads Gopkg.lock file
directly to figure out the base versions of all dependencies, and uses
usual logic to figure out latest versions of each dependency. This in
turn bypasses the constraints set in Gopkg.toml file.

That is not acceptable because it's not helpful to the user to see
these updates without distinction of what is within the Gopkg.toml
constraints and without. Also, the update button will do the wrong
thing, as it updates via dep ensure -update which does respect
Gopkg.toml constraints.

Helps #81.