Modify -dep flag behavior to use dep status to determine updates.
This change implements correct behavior, where the displayed updates
are within the Gopkg.toml constraints as determined by dep status
command.

We defer the work of determining remotely available updates to dep,
since there's no easy way to perform constraint checking on our end.
For now, this is a simple solution that works. In the future, a more
elaborate setup can be considered, if needed to have better performance
or implement certain additional features.

Resolves #81.