Support updates in dep projects (within Gopkg.toml constraints). (#82)
This change adds support for dep. More precisely, it adds support for
visualizing and updating dependencies (within Gopkg.toml constraints)
in dep-enabled projects.

A new -dep flag is added. Using it requires dep binary to be available
in PATH. When -dep flag is set to point to a Gopkg.toml file of a dep-
enabled project, it runs `dep status` in the containing directory to
determine available updates (within Gopkg.toml constraints) and then
displays them in a visual form.

The Update (and Update All) buttons are functional. Pressing them
executes `dep ensure -update {root}` command in same directory,
causing the update to be applied.

E.g.:

	$ Go-Package-Store -dep=/Users/Gopher/go/src/github.com/bradleyfalzon/gopherci/Gopkg.toml
	Determining the list of Go packages from Gopkg.toml file: /Users/Gopher/go/src/github.com/bradleyfalzon/gopherci/Gopkg.toml
	Go Package Store server is running at http://localhost:7043/updates.

	dep ensure -update github.com/google/go-github
	Done.

Resolves #81.