presenter: Create Repo type with fields only for presenters.
Previously, the top-level gps.Repo type was being given as input to
presenters.

The problem is that it contains multiple overlapping fields that are
used during workspace processing. This makes it hard for workspace to
know which fields need to be set, and makes it hard for presenters to
know which fields they can/should use.

Create a much smaller, well defined presenter.Repo type that contains
only the needed input for presenters. This makes everything cleaner.

This is a refactor only, there should not be a change in behavior.
Resolves the following TODO:

	// TODO: Right now, all presenters use Remote.RepoURL as the canonical remote URL, so it must be
	//       always set. This is a little confusing and redundant (since there's also Local.RemoteURL
	//       and just RemoteURL). Should change it so there's only one canonical remote URL for
	//       presenters to use.