Move "Recently Installed Updates" section to the top.
This change implements a minor presentation redesign, where the
"Recently Installed Updates" section is moved at the top, above
the "Updates Available" section. Previously it was at the bottom.

The motivation to do this is to improve the user experience when
installing updates from top to bottom. This way, the recently
updated package moves up to the section above, which is a much
smaller jump than moving it all the way to the bottom of the page.

Clean up some of the code in the process.

Regenerate assets.

Fixes #63.